VB-4974 API updates and rename 'visits' to 'visitsByDate' (#912)
